% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsiniw.R
\name{dsiniw}
\alias{dsiniw}
\title{The density function of the SinInverseWeibull probability distribution.}
\usage{
dsiniw(x, alpha, theta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
The density function of the SinInverseWeibull probability distribution.
}
\examples{
dsiniw(0.5,1,1)
dsiniw(0.5,0.5,0.7)
}

