% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsiniw.R
\name{hsiniw}
\alias{hsiniw}
\title{The hazard rate function of the SinInverseWeibull probability distribution.}
\usage{
hsiniw(x, alpha, theta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
The hazard rate function of the SinInverseWeibull probability distribution.
}
\examples{
hsiniw(0.5,0.5,1.1)
hsiniw(0.5,1,1.9)
}

