% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiniw.R
\name{psiniw}
\alias{psiniw}
\title{The cumulative function of the SinInverseWeibull probability distribution.}
\usage{
psiniw(q, alpha, theta, lower = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}

\item{lower}{Lower parameter.}

\item{log.p}{Log.p parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
The cumulative function of the SinInverseWeibull probability distribution.
}
\examples{
psiniw(0.5,1,1,TRUE,FALSE)
psiniw(0.5,0.5,0.7,TRUE,FALSE)
}

