% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsiniw.R
\name{qsiniw}
\alias{qsiniw}
\title{Te quantile function of the SinInverseWeibull probability distribution.}
\usage{
qsiniw(p, alpha, theta, lower = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}

\item{lower}{Lower parameter.}

\item{log.p}{Log.p parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
Te quantile function of the SinInverseWeibull probability distribution.
}
\examples{
qsiniw(0.5,1,1,TRUE,FALSE)
qsiniw(0.5,1,0.1,TRUE,FALSE)
}

