% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsiniw.R
\name{rsiniw}
\alias{rsiniw}
\title{Generates random deviates from a SinInverseWeibull probability distribution.}
\usage{
rsiniw(n, alpha, theta)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
Generates random deviates from a SinInverseWeibull probability distribution.
}
\examples{
rsiniw(1000,0.1,0.9)
rsiniw(1000,0.2,0.8)
}

