% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssiniw.R
\name{ssiniw}
\alias{ssiniw}
\title{The survival function of the SinInverseWeibull probability distribution.}
\usage{
ssiniw(x, alpha, theta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{Alpha parameter.}

\item{theta}{Theta parameter.}
}
\value{
A vector with n observations of the SinInverseWeibull distribution.
}
\description{
The survival function of the SinInverseWeibull probability distribution.
}
\examples{
ssiniw(0.1, 1, 1)
ssiniw(0.1, 1, 0.1)
}

